+++ docs/Changelog.md
@@ -7,7 +7,9 @@
 - Fractional zoom [#75](https://github.com/mapsforge/mapsforge/issues/75)
     - `Parameters.FRACTIONAL_ZOOM`
 - Motorider map theme [#1483](https://github.com/mapsforge/mapsforge/issues/1483)
-- Fix hillshading on 0 lat / lon [#1497](https://github.com/mapsforge/mapsforge/issues/1497)
+- Fix hillshading at 0 lat / lon [#1497](https://github.com/mapsforge/mapsforge/issues/1497)
+- Polygon rendering zoom option [#1503](https://github.com/mapsforge/mapsforge/pull/1503)
+  - `Parameters.POLYGON_ZOOM_MIN`
 - `mapsforge-themes` change package [#1135](https://github.com/mapsforge/mapsforge/issues/1135)
   - Rename `InternalRenderTheme` to `MapsforgeThemes`
 - Minor improvements and bug fixes